+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [4.2.0] - 2-19-2024
+
+### Added
+- LSP client and server support to optionally send and receive insertTextFormat completion item parameter
+- LSP server check for client snippet support capability to pick between plain text or snippet completion
+
+### Fixed
+- LSP server to handle multiple documents rather than previous limitation of one input per server process
+
 ## [4.1.0] - 2-06-2024
 
 ### Added